Move csv array transformation
